Implemented requested changes by @qvalentin (#9)
