+++ src/test/resources/UserAuthorization.feature
@@ -9,10 +9,10 @@ Background:
 
 Scenario: Successful login with username and password.
   When user requests login with username "user" and password "secure_password"
-  And response status code is 200
+  Then response status code is 200
   And response contains refreshToken "token" and the user with id 1234
 
-Scenario: Failed login with username and password.
+Scenario: Failed login with wrong username or password.
   When user requests login with username "user" and password "wrong_password"
   Then response contains key "message" and value "User could not be authenticated. No User found for username or password."
   And response contains key "status" and value "denied"
@@ -24,12 +24,20 @@ Scenario: Successful creation of new accessToken with refreshToken.
   And response contains valid accessToken for user 1234
   And response status code is 200
 
-Scenario: Successful request of existing accessToken with refreshToken.
+Scenario: Successful retrieval of existing accessToken with refreshToken.
   Given accessToken with value "6bb9cb4f-7b51-4c0a-8013-ed7a34e56282" exists for user 1234
   When user requests accessToken with refreshToken "token" and userId 1234
   Then response contains key "userId" and value "1234"
   And response contains valid accessToken for user 1234
   And response status code is 200
+
+Scenario: Successful retrieval of freshly created accessToken with refreshToken
+  Given accessToken with value "6bb9cb4f-7b51-4c0a-8013-ed7a34e56282" exists for user 1234 and is valid until 0
+  When user requests accessToken with refreshToken "token" and userId 1234
+  Then response contains key "userId" and value "1234"
+  And response contains valid accessToken for user 1234 with a different value than "6bb9cb4f-7b51-4c0a-8013-ed7a34e56282"
+  And response status code is 200
+
 
   Scenario: Failed retrieval of accessToken with wrong refreshToken.
   When user requests accessToken with refreshToken "not_the_token" and userId 1234
